Add fixtures
