+++ LoopTests/Fixtures/basal_profile.json
@@ -1,0 +1,44 @@
+[
+  {
+    "i": 0,
+    "start": "00:00:00",
+    "rate": 0.9,
+    "minutes": 0
+  },
+  {
+    "i": 1,
+    "start": "04:00:00",
+    "rate": 0.925,
+    "minutes": 240
+  },
+  {
+    "i": 2,
+    "start": "07:00:00",
+    "rate": 0.85,
+    "minutes": 420
+  },
+  {
+    "i": 3,
+    "start": "10:00:00",
+    "rate": 0.85,
+    "minutes": 600
+  },
+  {
+    "i": 4,
+    "start": "12:00:00",
+    "rate": 0.75,
+    "minutes": 720
+  },
+  {
+    "i": 5,
+    "start": "15:00:00",
+    "rate": 0.8,
+    "minutes": 900
+  },
+  {
+    "i": 6,
+    "start": "22:00:00",
+    "rate": 0.9,
+    "minutes": 1320
+  }
+]
