+++ LoopTests/Fixtures/counteraction_effect_falling_glucose.json
@@ -1,0 +1,20 @@
+[
+  {
+    "value" : -4.5999999999999988,
+    "startDate" : "2015-10-25T19:15:00",
+    "unit" : "mg\/min·dL",
+    "endDate" : "2015-10-25T19:20:00"
+  },
+  {
+    "value" : -2.5999999999999996,
+    "startDate" : "2015-10-25T19:20:00",
+    "unit" : "mg\/min·dL",
+    "endDate" : "2015-10-25T19:25:00"
+  },
+  {
+    "value" : -0.59999999999999998,
+    "startDate" : "2015-10-25T19:25:00",
+    "unit" : "mg\/min·dL",
+    "endDate" : "2015-10-25T19:30:00"
+  }
+]
