+++ LoopTests/Fixtures/dynamic_glucose_effect_partially_observed.json
@@ -1,0 +1,372 @@
+[
+  {
+    "amount" : 0,
+    "unit" : "mg\/dL",
+    "date" : "2015-10-15T21:30:00"
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-15T21:35:00",
+    "amount" : 0
+  },
+  {
+    "unit" : "mg\/dL",
+    "amount" : 0,
+    "date" : "2015-10-15T21:40:00"
+  },
+  {
+    "amount" : 25.000000000000004,
+    "date" : "2015-10-15T21:45:00",
+    "unit" : "mg\/dL"
+  },
+  {
+    "date" : "2015-10-15T21:50:00",
+    "amount" : 40,
+    "unit" : "mg\/dL"
+  },
+  {
+    "amount" : 50.000000000000007,
+    "unit" : "mg\/dL",
+    "date" : "2015-10-15T21:55:00"
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-15T22:00:00",
+    "amount" : 65
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-15T22:05:00",
+    "amount" : 90
+  },
+  {
+    "date" : "2015-10-15T22:10:00",
+    "amount" : 94.399999999999991,
+    "unit" : "mg\/dL"
+  },
+  {
+    "unit" : "mg\/dL",
+    "amount" : 98.983333333333334,
+    "date" : "2015-10-15T22:15:00"
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-15T22:20:00",
+    "amount" : 103.56666666666668
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-15T22:25:00",
+    "amount" : 108.15000000000001
+  },
+  {
+    "date" : "2015-10-15T22:30:00",
+    "unit" : "mg\/dL",
+    "amount" : 112.73333333333333
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-15T22:35:00",
+    "amount" : 117.31666666666668
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-15T22:40:00",
+    "amount" : 121.90000000000001
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-15T22:45:00",
+    "amount" : 126.48333333333333
+  },
+  {
+    "date" : "2015-10-15T22:50:00",
+    "amount" : 131.06666666666666,
+    "unit" : "mg\/dL"
+  },
+  {
+    "amount" : 135.65000000000001,
+    "unit" : "mg\/dL",
+    "date" : "2015-10-15T22:55:00"
+  },
+  {
+    "date" : "2015-10-15T23:00:00",
+    "unit" : "mg\/dL",
+    "amount" : 140.23333333333335
+  },
+  {
+    "amount" : 144.81666666666666,
+    "unit" : "mg\/dL",
+    "date" : "2015-10-15T23:05:00"
+  },
+  {
+    "amount" : 149.40000000000001,
+    "date" : "2015-10-15T23:10:00",
+    "unit" : "mg\/dL"
+  },
+  {
+    "date" : "2015-10-15T23:15:00",
+    "amount" : 153.98333333333335,
+    "unit" : "mg\/dL"
+  },
+  {
+    "date" : "2015-10-15T23:20:00",
+    "amount" : 158.56666666666666,
+    "unit" : "mg\/dL"
+  },
+  {
+    "date" : "2015-10-15T23:25:00",
+    "unit" : "mg\/dL",
+    "amount" : 163.15000000000001
+  },
+  {
+    "unit" : "mg\/dL",
+    "amount" : 167.73333333333335,
+    "date" : "2015-10-15T23:30:00"
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-15T23:35:00",
+    "amount" : 172.31666666666669
+  },
+  {
+    "date" : "2015-10-15T23:40:00",
+    "unit" : "mg\/dL",
+    "amount" : 176.90000000000001
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-15T23:45:00",
+    "amount" : 181.48333333333335
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-15T23:50:00",
+    "amount" : 186.06666666666669
+  },
+  {
+    "amount" : 190.65000000000001,
+    "date" : "2015-10-15T23:55:00",
+    "unit" : "mg\/dL"
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T00:00:00",
+    "amount" : 195.23333333333335
+  },
+  {
+    "date" : "2015-10-16T00:05:00",
+    "amount" : 199.81666666666669,
+    "unit" : "mg\/dL"
+  },
+  {
+    "date" : "2015-10-16T00:10:00",
+    "amount" : 204.40000000000001,
+    "unit" : "mg\/dL"
+  },
+  {
+    "date" : "2015-10-16T00:15:00",
+    "amount" : 208.98333333333335,
+    "unit" : "mg\/dL"
+  },
+  {
+    "unit" : "mg\/dL",
+    "amount" : 213.56666666666669,
+    "date" : "2015-10-16T00:20:00"
+  },
+  {
+    "amount" : 218.15000000000001,
+    "date" : "2015-10-16T00:25:00",
+    "unit" : "mg\/dL"
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T00:30:00",
+    "amount" : 220
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T00:35:00",
+    "amount" : 220
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T00:40:00",
+    "amount" : 220
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T00:45:00",
+    "amount" : 220
+  },
+  {
+    "date" : "2015-10-16T00:50:00",
+    "amount" : 220,
+    "unit" : "mg\/dL"
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T00:55:00",
+    "amount" : 220
+  },
+  {
+    "amount" : 220,
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T01:00:00"
+  },
+  {
+    "amount" : 220,
+    "date" : "2015-10-16T01:05:00",
+    "unit" : "mg\/dL"
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T01:10:00",
+    "amount" : 220
+  },
+  {
+    "unit" : "mg\/dL",
+    "amount" : 220,
+    "date" : "2015-10-16T01:15:00"
+  },
+  {
+    "date" : "2015-10-16T01:20:00",
+    "amount" : 220,
+    "unit" : "mg\/dL"
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T01:25:00",
+    "amount" : 220
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T01:30:00",
+    "amount" : 220
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T01:35:00",
+    "amount" : 220
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T01:40:00",
+    "amount" : 220
+  },
+  {
+    "date" : "2015-10-16T01:45:00",
+    "amount" : 220,
+    "unit" : "mg\/dL"
+  },
+  {
+    "amount" : 220,
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T01:50:00"
+  },
+  {
+    "amount" : 220,
+    "date" : "2015-10-16T01:55:00",
+    "unit" : "mg\/dL"
+  },
+  {
+    "amount" : 220,
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T02:00:00"
+  },
+  {
+    "amount" : 220,
+    "date" : "2015-10-16T02:05:00",
+    "unit" : "mg\/dL"
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T02:10:00",
+    "amount" : 220
+  },
+  {
+    "unit" : "mg\/dL",
+    "amount" : 220,
+    "date" : "2015-10-16T02:15:00"
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T02:20:00",
+    "amount" : 220
+  },
+  {
+    "amount" : 220,
+    "date" : "2015-10-16T02:25:00",
+    "unit" : "mg\/dL"
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T02:30:00",
+    "amount" : 220
+  },
+  {
+    "amount" : 220,
+    "date" : "2015-10-16T02:35:00",
+    "unit" : "mg\/dL"
+  },
+  {
+    "date" : "2015-10-16T02:40:00",
+    "unit" : "mg\/dL",
+    "amount" : 220
+  },
+  {
+    "date" : "2015-10-16T02:45:00",
+    "amount" : 220,
+    "unit" : "mg\/dL"
+  },
+  {
+    "date" : "2015-10-16T02:50:00",
+    "amount" : 220,
+    "unit" : "mg\/dL"
+  },
+  {
+    "date" : "2015-10-16T02:55:00",
+    "unit" : "mg\/dL",
+    "amount" : 220
+  },
+  {
+    "date" : "2015-10-16T03:00:00",
+    "unit" : "mg\/dL",
+    "amount" : 220
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T03:05:00",
+    "amount" : 220
+  },
+  {
+    "unit" : "mg\/dL",
+    "amount" : 220,
+    "date" : "2015-10-16T03:10:00"
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T03:15:00",
+    "amount" : 220
+  },
+  {
+    "amount" : 220,
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T03:20:00"
+  },
+  {
+    "unit" : "mg\/dL",
+    "amount" : 220,
+    "date" : "2015-10-16T03:25:00"
+  },
+  {
+    "amount" : 220,
+    "date" : "2015-10-16T03:30:00",
+    "unit" : "mg\/dL"
+  },
+  {
+    "unit" : "mg\/dL",
+    "date" : "2015-10-16T03:35:00",
+    "amount" : 220
+  }
+]
