+++ LoopTests/Fixtures/momentum_effect_bouncing.json
@@ -1,0 +1,42 @@
+[
+  {
+    "date": "2015-10-25T19:25:00",
+    "amount": 0.0,
+    "unit": "mg/dL"
+  },
+  {
+    "date": "2015-10-25T19:30:00",
+    "amount": 0.23051025736941719,
+    "unit": "mg/dL"
+  },
+  {
+    "date": "2015-10-25T19:35:00",
+    "amount": 3.2371657882748588,
+    "unit": "mg/dL"
+  },
+  {
+    "date": "2015-10-25T19:40:00",
+    "amount": 6.2438213191803005,
+    "unit": "mg/dL"
+  },
+  {
+    "date": "2015-10-25T19:45:00",
+    "amount": 9.2504768500857413,
+    "unit": "mg/dL"
+  },
+  {
+    "date": "2015-10-25T19:50:00",
+    "amount": 12.257132380991184,
+    "unit": "mg/dL"
+  },
+  {
+    "date": "2015-10-25T19:55:00",
+    "amount": 15.263787911896625,
+    "unit": "mg/dL"
+  },
+  {
+    "date": "2015-10-25T20:00:00",
+    "amount": 18.270443442802062,
+    "unit": "mg/dL"
+  }
+]
